requirements.txt
streamlit
transformers
sentencepiece
langdetect
pytesseract
PyMuPDF
pillow
